requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
exceptiongroup==1.1.1
Flask==2.3.2
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
parsel==1.8.1
pefile==2023.2.7
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
pyparsing==3.1.0
PySocks==1.7.1
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.3
undetected-chromedriver==3.5.0
uritemplate==4.1.1
urllib3==1.26.16
w3lib==2.1.1
websockets==11.0.3
Werkzeug==2.3.6
wsproto==1.2.0
